+++ frontend/src/pages/SignUp/Agreements.tsx
