+++ frontend/src/pages/SignUp/DogDetailInfo.tsx
